Add pairing completion modal
Keep track of pairingRequest status in redux.

For now, redux state is the only driver of the pairing UI. I included
url-driven modals originally, thinking that would be the easiest way to
trigger pairing UI while the main window isn't open (e.g., through an OS
notification). TBD if that's actually a requirement; and if so, what the
best approach will be now.

A device list is shown in the settings screen for development.