Use name digest as protocol ID
Name is considered globally unique for a protocol (as far as any user is
concerned). This provides an ID that is consistent across Server
intsalls, so that data may be exported to a different server than the
protocol was imported from.

This also clarifies some error messages (MissingProtocol is
specifically for a missing container file) and adds test coverage.