Update web browser support
- Fixes some JS errors present when building for the web target.
- Add new required shims
- Send API_INFO to app container

This allows the start screen to load in a dev browser (outside
electron).

However, this workflow is decreasingly useful, as IPC is needed for
most functionality. I've removed documentation for browser development
support. With support for electron:dev and chrome devtools within
electron, I don't see a strong use case for browser development and may
remove it entirely in the future.